Fixed Activity Log, Number inputs and Invoice Edit

+++ database/migrations/2017_06_27_142308_create_vats_table.php
@@ -15,7 +15,7 @@ class CreateVatsTable extends Migration
         Schema::create('vats', function (Blueprint $table) {
             $table->increments('id');
             $table->string('vat_code')->unique();
-            $table->integer('rate');
+            $table->decimal('rate');
             $table->string('description');
             $table->softDeletes();
             $table->timestamps();
